practice constructors, refactor and optimization all clases

+++ src/com/alurapractice/screenmatch/calculos/Review.java
@@ -1,5 +1,5 @@
 package com.alurapractice.screenmatch.calculos;
 
 public interface Review {
-    int getReview();
+    double getReview();
 }
